sépare les tests d'intégration en plusieurs fichiers
